mentorPlanilha/modelo_config.py
# Pacotes necessários
from langchain_groq import ChatGroq
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser
from langchain_core.chat_history import InMemoryChatMessageHistory
from langchain_core.runnables.history import RunnableWithMessageHistory
from dotenv import load_dotenv
import os

load_dotenv()
api_key = os.getenv("GROQ_API_KEY")

modelo = ChatGroq(
    model="llama-3.3-70b-versatile",
    temperature=0.7
)
