mentorPlanilha/pyproject.toml
[project]
name = "mentorplanilha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.3",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "python-dotenv>=1.2.1",
]
